+++ src/Viewer/LiteMol-viewer.js
@@ -7,7 +7,7 @@ var LiteMol;
 (function (LiteMol) {
     var Viewer;
     (function (Viewer) {
-        Viewer.VERSION = { number: "1.1.9", date: "Nov 18 2016" };
+        Viewer.VERSION = { number: "1.1.10", date: "Nov 22 2016" };
     })(Viewer = LiteMol.Viewer || (LiteMol.Viewer = {}));
 })(LiteMol || (LiteMol = {}));
 /*